+++ cloud/functions/src/model/data/slide.ts
