+++ cloud/functions/src/request/utils/api-utils.ts
@@ -2,7 +2,7 @@ import * as functions from 'firebase-functions';
 
 import fetch, {Response} from 'node-fetch';
 
-import {Deck} from '../../model/deck';
+import {Deck} from '../../model/data/deck';
 import {ApiDeck} from '../../model/api/api.deck';
 import {ApiPresentation} from '../../model/api/api.presentation';
