+++ cloud/functions/src/utils/resources/resources.ts
