+++ cloud/functions/src/watch/delete/utils/delete-slides-utils.ts
@@ -1,6 +1,6 @@
 import * as admin from 'firebase-admin';
 
-import {Slide} from '../../../model/slide';
+import {Slide} from '../../../model/data/slide';
 
 export interface DeckSlides {
   deckId: string;
